fix: INCONSISTENT_IMPORT_ASSERTIONS warning
Rollup v3 has introduced some support of import assertions in
https://github.com/rollup/rollup/pull/4646/. In particular, it uses
an `{ assertions: { type: <xxx> } }` object to track the assertions
linked to a module. However, this does not seem to be set in preload,
whereas it _is_ set afterwards. This results in a warning.

In order to silence the warning, we need to explicitly tell rollup
about the assertions type, so that this piece of information is
available even at preload time.

Note: this also makes the test 'handles garbage' pass. Without this
rollup issues 2 warnings instead of 1 expected (resulting in test failure).